Extract JavaVersionPredicates
