+++ junit-jupiter-engine/src/test/java/org/junit/jupiter/api/condition/EnabledOnJreConditionTests.java
@@ -12,22 +12,22 @@ package org.junit.jupiter.api.condition;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava10;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava11;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava12;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava13;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava14;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava15;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava16;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava17;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava18;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava19;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava20;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava21;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava22;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava23;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava8;
-import static org.junit.jupiter.api.condition.EnabledOnJreIntegrationTests.onJava9;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava10;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava11;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava12;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava13;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava14;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava15;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava16;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava17;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava18;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava19;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava20;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava21;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava22;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava23;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava8;
+import static org.junit.jupiter.api.condition.JavaVersionPredicates.onJava9;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExecutionCondition;
